+++ x/pylons/client/cli/easel_test.go
@@ -86,9 +86,7 @@ func createEaselCookbook(t *testing.T, simInfo *easelBasicSim) {
 }
 
 func createMintRecipe1(t *testing.T, simInfo *easelBasicSim) {
-	coinInputs, err := json.Marshal([]types.CoinInput{
-		{Coins: sdk.NewCoins(sdk.NewCoin("node0token", sdk.NewInt(10)))},
-	})
+	coinInputs, err := json.Marshal([]string{"10node0token", "10uatom,10upylon"})
 	require.NoError(t, err)
 
 	entries, err := json.Marshal(types.EntriesList{
@@ -128,7 +126,7 @@ func createMintRecipe1(t *testing.T, simInfo *easelBasicSim) {
 		string(entries),
 		string(itemOutputs),
 		"0",
-		"{\"denom\": \"upylon\", \"amount\": \"12\"}",
+		"12upylon",
 		"true",
 		"extraInfo",
 	}
@@ -149,7 +147,7 @@ func mintNFT1(t *testing.T, simInfo *easelBasicSim) {
 	out, err := clitestutil.ExecTestCLICmd(simInfo.ctx, cli.CmdExecuteRecipe(), args)
 	require.NoError(t, err)
 	var resp sdk.TxResponse
-	require.NoError(t, simInfo.ctx.JSONCodec.UnmarshalJSON(out.Bytes(), &resp))
+	require.NoError(t, simInfo.ctx.Codec.UnmarshalJSON(out.Bytes(), &resp))
 	require.Equal(t, uint32(0), resp.Code)
 
 	// simulate waiting for later block heights
@@ -167,7 +165,7 @@ func mintNFT1(t *testing.T, simInfo *easelBasicSim) {
 	out, err = clitestutil.ExecTestCLICmd(simInfo.ctx, cli.CmdShowExecution(), args)
 	require.NoError(t, err)
 	var execResp types.QueryGetExecutionResponse
-	require.NoError(t, simInfo.ctx.JSONCodec.UnmarshalJSON(out.Bytes(), &execResp))
+	require.NoError(t, simInfo.ctx.Codec.UnmarshalJSON(out.Bytes(), &execResp))
 	// verify completed
 	require.Equal(t, true, execResp.Completed)
 
@@ -178,7 +176,7 @@ func mintNFT1(t *testing.T, simInfo *easelBasicSim) {
 	out, err = clitestutil.ExecTestCLICmd(simInfo.ctx, cli.CmdShowItem(), args)
 	require.NoError(t, err)
 	var itemResp types.QueryGetItemResponse
-	require.NoError(t, simInfo.ctx.JSONCodec.UnmarshalJSON(out.Bytes(), &itemResp))
+	require.NoError(t, simInfo.ctx.Codec.UnmarshalJSON(out.Bytes(), &itemResp))
 	require.Equal(t, cookbookIDEasel, itemResp.Item.CookbookID)
 }
 
@@ -322,7 +320,7 @@ func mintNFT2(t *testing.T, simInfo *easelBasicSim) {
 	out, err := clitestutil.ExecTestCLICmd(simInfo.ctx, cli.CmdExecuteRecipe(), args)
 	require.NoError(t, err)
 	var resp sdk.TxResponse
-	require.NoError(t, simInfo.ctx.JSONCodec.UnmarshalJSON(out.Bytes(), &resp))
+	require.NoError(t, simInfo.ctx.Codec.UnmarshalJSON(out.Bytes(), &resp))
 	require.Equal(t, uint32(0), resp.Code)
 
 	// simulate waiting for later block heights
@@ -340,7 +338,7 @@ func mintNFT2(t *testing.T, simInfo *easelBasicSim) {
 	out, err = clitestutil.ExecTestCLICmd(simInfo.ctx, cli.CmdShowExecution(), args)
 	require.NoError(t, err)
 	var execResp types.QueryGetExecutionResponse
-	require.NoError(t, simInfo.ctx.JSONCodec.UnmarshalJSON(out.Bytes(), &execResp))
+	require.NoError(t, simInfo.ctx.Codec.UnmarshalJSON(out.Bytes(), &execResp))
 	// verify completed
 	require.Equal(t, true, execResp.Completed)
 
@@ -351,7 +349,7 @@ func mintNFT2(t *testing.T, simInfo *easelBasicSim) {
 	out, err = clitestutil.ExecTestCLICmd(simInfo.ctx, cli.CmdShowItem(), args)
 	require.NoError(t, err)
 	var itemResp types.QueryGetItemResponse
-	require.NoError(t, simInfo.ctx.JSONCodec.UnmarshalJSON(out.Bytes(), &itemResp))
+	require.NoError(t, simInfo.ctx.Codec.UnmarshalJSON(out.Bytes(), &itemResp))
 	require.Equal(t, cookbookIDEasel, itemResp.Item.CookbookID)
 
 	args = []string{cookbookIDEasel, simInfo.mintRecipeID2}